requirements.txt
google-generativeai
finance-datareader
python-telegram-bot
requests
python-dotenv
schedule
beautifulsoup4
plotly
